+++ apps/wizard/app_pages/dataset_preview/explore.py
@@ -10,8 +10,9 @@ from streamlit_agraph import Config, ConfigBuilder, Edge, Node, agraph
 
 from apps.wizard.utils import metadata_export_basic, set_states
 from etl.config import ENV_IS_REMOTE
+from etl.dag_utils import load_dag
 from etl.paths import DATA_DIR
-from etl.steps import extract_step_attributes, filter_to_subgraph, load_dag
+from etl.steps import extract_step_attributes, filter_to_subgraph
 
 # CONFIG
 st.set_page_config(
