+++ apps/wizard/etl_steps/utils.py
@@ -3,9 +3,9 @@ import os
 import streamlit as st
 
 from apps.wizard.utils import WIZARD_DIR
+from etl.dag_utils import load_dag
 from etl.helpers import read_json_schema
 from etl.paths import DAG_DIR, SCHEMAS_DIR
-from etl.steps import load_dag
 
 # Step icons
 STEP_ICONS = {
