+++ apps/wizard/utils/__init__.py
@@ -32,6 +32,7 @@ from wfork_streamlit_profiler import Profiler
 from apps.wizard.utils.defaults import load_wizard_defaults, update_wizard_defaults_from_form
 from apps.wizard.utils.step_form import StepForm
 from etl.config import OWID_ENV, SENTRY_DSN, enable_sentry
+from etl.dag_utils import load_dag
 from etl.db import read_sql
 from etl.files import ruamel_dump, ruamel_load
 from etl.metadata_export import main as metadata_export
@@ -44,7 +45,6 @@ from etl.paths import (
     STEPS_GRAPHER_DIR,
     STEPS_MEADOW_DIR,
 )
-from etl.steps import load_dag
 
 __all__ = [
     "load_wizard_defaults",
