+++ etl/steps/__init__.py
@@ -20,14 +20,13 @@ from dataclasses import dataclass, field
 from glob import glob
 from importlib import import_module
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, Optional, Protocol, Set, Union, cast
+from typing import Any, Dict, Iterable, List, Optional, Protocol, Set, cast
 from urllib.parse import urlparse
 
 import fasteners
 import pandas as pd
 import requests
 import structlog
-import yaml
 from owid import catalog
 from owid.catalog import s3_utils
 from owid.catalog.catalogs import OWID_CATALOG_URI
@@ -151,42 +150,6 @@ def traverse(graph: Graph, nodes: Set[str]) -> Graph:
         to_visit = to_visit.union(reachable[node])
 
     return dict(reachable)
-
-
-def load_dag(filename: Union[str, Path] = paths.DEFAULT_DAG_FILE) -> Graph:
-    return _load_dag(filename, {})
-
-
-def _load_dag(filename: Union[str, Path], prev_dag: Dict[str, Any]):
-    """
-    Recursive helper to 1) load a dag itself, and 2) load any sub-dags
-    included in the dag via 'include' statements
-    """
-    dag_yml = _load_dag_yaml(str(filename))
-    curr_dag = _parse_dag_yaml(dag_yml)
-
-    # make sure there are no fast-track steps in the DAG
-    if "fasttrack.yml" not in str(filename):
-        fast_track_steps = {step for step in curr_dag if "/fasttrack/" in step}
-        if fast_track_steps:
-            raise ValueError(f"Fast-track steps detected in DAG {filename}: {fast_track_steps}")
-
-    duplicate_steps = prev_dag.keys() & curr_dag.keys()
-    if duplicate_steps:
-        raise ValueError(f"Duplicate steps detected in DAG {filename}: {duplicate_steps}")
-
-    curr_dag.update(prev_dag)
-
-    for sub_dag_filename in dag_yml.get("include", []):
-        sub_dag = _load_dag(paths.BASE_DIR / sub_dag_filename, curr_dag)
-        curr_dag.update(sub_dag)
-
-    return curr_dag
-
-
-def _load_dag_yaml(filename: str) -> Dict[str, Any]:
-    with open(filename) as istream:
-        return yaml.safe_load(istream)
 
 
 def _parse_dag_yaml(dag: Dict[str, Any]) -> Dict[str, Any]:
