+++ etl/version_tracker.py
@@ -13,9 +13,10 @@ from rich_click.rich_command import RichCommand
 
 from etl import paths
 from etl.config import ADMIN_HOST
+from etl.dag_utils import load_dag
 from etl.db import can_connect
 from etl.grapher.io import get_info_for_etl_datasets
-from etl.steps import extract_step_attributes, load_dag, reverse_graph
+from etl.steps import extract_step_attributes, reverse_graph
 
 log = structlog.get_logger()
